Add duration and split simulation file
