+++ src/WrapperApp/components/Simulation/SimulationCard.tsx
@@ -37,6 +37,7 @@ import {
 	SimulationInputFiles,
 	StatusState
 } from '../../../types/ResponseTypes';
+import { millisecondsToTimeString } from '../../../util/time';
 import { SimulationProgressBar } from './SimulationProgressBar';
 
 type SimulationCardProps = {
@@ -170,6 +171,12 @@ export default function SimulationCard({
 		loadFromJson(inputJson);
 	};
 
+	const { startTime, endTime } = simulationStatus;
+
+	const startDate = new Date(startTime);
+	const endDate = endTime ? new Date(endTime) : undefined;
+	const duration = endDate ? endDate.valueOf() - startDate.valueOf() : 0;
+
 	return (
 		<Card
 			sx={{ height: '100%', display: 'flex', flexDirection: 'column' }}
@@ -184,14 +191,12 @@ export default function SimulationCard({
 			/>
 			<CardHeader
 				title={`${simulationStatus.title}`}
-				subheader={`${simulationStatus.startTime.toLocaleString('en-US').split(' ')[0]} ${
-					simulationStatus.startTime.toLocaleString('en-US').split(' ')[4] ?? '00:00:00'
-				} - ${
-					currentJobStatusData[StatusState.COMPLETED](simulationStatus) &&
-					simulationStatus.endTime
-						? simulationStatus.endTime.toLocaleString('en-US').split(' ')[4] ?? '?'
-						: '?'
-				}`}
+				subheader={
+					<>
+						{`${startDate.toLocaleDateString()} ${startDate.toLocaleTimeString()}  `}
+						<br /> {endDate ? `Duration: ${millisecondsToTimeString(duration)}` : ''}
+					</>
+				}
 				action={
 					handleDelete &&
 					(currentJobStatusData[StatusState.COMPLETED](simulationStatus) ||
